feat(get) resurrect stale data on callback soft errors
Implement a new `opts.resurrect_ttl` instance and `get()` option. When
specified, if a callback lookup is triggered, but returns a soft error
(`nil, err`), we resurrect the stale value if it is still available for
the duration of the `resurrect_ttl`.

We assume such callback errors to be caused by an I/O issue (e.g.
timeout, connection refused...). When such errors happen, the previous
behaviour would not do its best effort to serve stale data (and ensure
resiliency).

When stale data is returned by `get()`, the `hit_lvl` return value is
now set to `4`. This return value does not indicate whether or not a
given value was resurrected, but only that it is stale data that was
served to ensure resiliency. This behavior may change in the future to
be `4` on all subsequent retrievals of stale values, which would be more
valuable for applications to monitor.

Resilient usage:

    local data, err, hit_lvl = cache:get(key, { resurrect_ttl = 10 }, cb)
    if err then
        if not hit_lvl then
            ngx.log(ngx.ERR, err)
            return
        end

        -- The callback errored, but we are lucky: the value
        -- was still in the L2 shm.
        --
        -- It has been resurrected for 10 seconds.
        -- Next :get() will return no error and hit_lvl of `2`.

        ngx.log(ngx.WARN, "callback failed but stale value available ",
                          "(", err, ")")
    end

    -- `data` is a cached value (it can still be nil for negative hits)